scraper/requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
anyio==3.6.2
async-timeout==4.0.2
attrs==22.1.0
autoflake==1.7.6
autopep8==1.7.0
black==22.10.0
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
elastic-transport==8.4.0
elasticsearch==8.4.2
frozenlist==1.3.1
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.4
isort==5.10.1
Jinja2==3.1.2
MarkupSafe==2.1.1
multidict==6.0.2
mypy-extensions==0.4.3
oauthlib==3.2.2
openapi-python-client==0.11.6
pathspec==0.10.1
platformdirs==2.5.2
pycodestyle==2.9.1
pydantic==1.10.2
pyflakes==2.5.0
python-dateutil==2.8.2
python-dotenv==0.21.0
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
shellingham==1.5.0
six==1.16.0
sniffio==1.3.0
toml==0.10.2
tomli==2.0.1
typer==0.6.1
typing_extensions==4.4.0
urllib3==1.26.12
yarl==1.8.1
